[feat] 헤더 재사용
